backend_fastapi/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-dotenv
pydantic
pydantic-settings
